* Changed result type of `query.Session.NodeID()` from `int64` to `uint32` for compatibility with table session and discovery

+++ query/session.go
@@ -13,7 +13,7 @@ import (
 type (
 	SessionInfo interface {
 		ID() string
-		NodeID() int64
+		NodeID() uint32
 		Status() string
 	}
